clearer/more consistent thunk type names
some things were XEnum, some things were RunX; now it's all ThunkX

also moved them out into a separate file in anticipation of making a
Thunk a Value, so it can have more breathing room
